Duplicate Modules: Add option to include their bytes
Two bundles may have large byte differences mainly because of some
modules being duplicated more or less often.

The totals show if the whole bundle has grown or shrunk but the modules
section couldn't tell you which modules had increased or decreased the
number of duplicates. Thus you could know the change was because of
altered duplication, but it would be hard to identify which modules
caused the change.

This hooks up the ModuleMetric selector to include the byte size
of the duplicates when comparing module sizes.

There are other ways of getting this done, but I tried to choose the
least invasive way. doing this operation as a selector allows pretty
easy extension to measuring other types (like duplicate count)