Module Duplication: list chunks from both Jobs
In the module info dialog we list the chunks that a module is in.
Previously, we only showed the list of chunks that it's in from *one* of
the jobs. Listing all chunks allows you to see if a module is
duplicated and how that duplication is different across jobs.